fixed transform not preserved with multiple animations
When multiple Animations were used, the preTransform property would "remember"
the second to last transform instead of the first.  Therefore, this restored
the correct transform preserving behavior only when one Animation was used.
Also, the transform is now restored via the UIView.animation closure capturing
a constant from the function context instead of adding a computed property
to UIView, which could interfere with subclasses using this property.

Test for this case added in example app's tests.